% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFT_fit.R
\name{AFT_fit}
\alias{AFT_fit}
\title{Fit Additive AFT Model}
\usage{
AFT_fit(
  data,
  mset,
  pset,
  lambda1,
  lambda2,
  lambda1s,
  lambda2s,
  penalty,
  M = 300,
  epsilon = 1e-04
)
}
\arguments{
\item{data}{a list containing:\tabular{ll}{
\code{y} \tab a numeric vector recording logarithm survival time. \cr
\tab \cr
\code{x} \tab the standardized covariates matrix with all entries between -1 and 1.\cr
\tab \cr
\code{delta} \tab a numeric vector that consists of status indicator, normally 0=alive, 1=dead. \cr
}}

\item{mset}{a vector, the number of intervals for each continuous variable to expand the B-spline.}

\item{pset}{a vector, the degree of the B-spline expanded for each continuous variable.}

\item{lambda1}{a number, penalty parameters for linear effects.}

\item{lambda2}{a number, penalty parameters for nonlinear effects.}

\item{lambda1s}{a vector, penalty parameters for linear effects, used to select the best hyperparameter.}

\item{lambda2s}{a vector, penalty parameters for nonlinear effects, used to select the best hyperparameter.}

\item{penalty}{a character, which indicates the penalty function, and can be one of "glasso", "gscad" and "gmcp".}

\item{M}{a number, the maximum times of iteration of BMD algorithm.}

\item{epsilon}{eps a number, the tolerance error.}
}
\value{
a list containing:\tabular{ll}{
\code{ny} \tab a vector, the logarithm survival time after sorting in ascending order.\cr
\tab \cr
\code{nx} \tab a vector, x sorted by rows corresponding to \code{ny}.\cr
\tab \cr
\code{BS} \tab a matrix, the columns of which should be the predicted value
of the orthogonal B-spline base at the corresponding point for continuous
variables or standardized data for discrete variables.\cr
\tab \cr
\code{theta_est} \tab a vector, the estimated values of parameters.\cr
\tab \cr
\code{best_lambda} \tab a vector, the best value for \code{lambda1} and \code{lambda2} according to EBIC criterion.\cr
\tab \cr
\code{group_re} \tab a list, the estimated values of parameters for each group.\cr
\tab \cr
\code{model_struct} \tab a vector, identify the effect type of each component.\cr}
}
\description{
This function fits an additive model for given standardized data with BMD algorithm.
}
\examples{
mset <- rep(4,30)
pset <- rep(3,30)
lam_gscad1 <- seq(0.03, 0.04, 0.002)
lam_gscad2 <- seq(0.03, 0.04, 0.002)
re <- AFT_fit(SimuData, mset, pset, lambda1s = lam_gscad1, lambda2s = lam_gscad2,
              penalty = "gscad", M = 300)
# coefficients estimate
re$theta_est
# best pair of regulation parameters
re$best_lambda
# model
re$model_structure
}
