% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR.R
\name{AR}
\alias{AR}
\title{Auto-Regression Time Series}
\usage{
AR(n, d, rho)
}
\arguments{
\item{n}{the number of rows of the result matrix, denotes the number of obversations.}

\item{d}{the number of columns of the result matrix, denotes the number of covariates.}

\item{rho}{the number to control the correlation between two adjacent columns of generated data.}
}
\value{
A matrix with \code{n} rows and \code{d} columns.
}
\description{
This function generates data for auto-regression time series.
}
\details{
The result matrix is a design matrix and the rows
denote observations and the columns denote covariates. The covariates
generated from a \code{d}-dimension normal distribution with mean zero
and covariance matrix \eqn{\Sigma = (\delta_{ij})},
where \eqn{\delta_{ij} = \rho^{i-j}}. But since we controled
the range between (-1, 1), result obtained by using \code{\link[stats]{cov}}
is not approximately \eqn{\rho}.
}
\examples{
AR(200,30,0.25)

}
