% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_bs.R
\name{bar_bs}
\alias{bar_bs}
\title{Calculate Orthogonal B-spline Basis.}
\usage{
bar_bs(bs, d)
}
\arguments{
\item{bs}{a matrix, each column of which is the predicted value of the B-spline base at the corresponding point and the B-splines can be generated by \code{\link[splines2]{bSpline}} in \code{splines2} package.}

\item{d}{the number of continuous variables in initial dataset.}
}
\value{
A matrix, each column of which is the predicted value of the orthogonal B-spline base at the corresponding point.
}
\description{
This function calculates orthogonal B-spline basis.
}
\details{
The result matrix is obtained by Gram-Schmidt method.
}
\examples{
y <- rnorm(100)
bs <- splines2::bSpline(y,degree = 5)
# assume that there are 6 continuous variables in initial dataset
bar_bs(bs, d = 6)
}
