% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_tau.R
\name{find_tau}
\alias{find_tau}
\title{Censoring Rate Control}
\usage{
find_tau(c.r, n, d, rho, fun_list, col_index, tol = 1e-05, N.max = 30)
}
\arguments{
\item{c.r}{a number, the censoring rate you want to control.}

\item{n}{the number of observations you want to generate.}

\item{d}{the number of covariates you want to generate.}

\item{rho}{the number to control the correlation between two adjacent columns of generated data.}

\item{fun_list}{a list, whose elements are functions to be applied to covariates specified by \code{col_index} one by one.}

\item{col_index}{a vector, which specifies the indexes of covariates corresponding to \code{fun_list} in the data matrix.}

\item{tol}{a number, the tolerance error of \code{c.r} when searching for \eqn{\tau}.}

\item{N.max}{a number, which is used to control the number of observations (\code{N.max} * \code{n}) for searching for \eqn{\tau}.}
}
\value{
a number  \eqn{\tau}  to control the censoring rate \code{c.r}.
}
\description{
This function finds a number \eqn{\tau} to control the censoring rate.
}
\details{
Assume the logarithm of the survival time \eqn{T_i} following an
additive accelerated failure time model:
\deqn{T_i = \sum_{j=1}^pg_j(Z_{ij})+\varepsilon_i,}
where \eqn{Z_{ij}} denotes the value of \eqn{jth} covariate of
\eqn{ith} individual. \code{col_index} specifies which covariates have
a functional effect on survival time while \code{fun_list} specifies this
functions. Assume the censoring time follow the uniform distribution
\eqn{U[0,\tau]}, This function is used to find \eqn{\tau} to control
the censoring rate by using bisection method.
}
\examples{
f1 <- function(x) return(x)
f2 <- function(x) return(-x)
f3 <- function(x) return(x^2)
find_tau(c.r = 0.5, n = 200, d = 30, rho = 0.25,
         fun_list = list(f1, f2, f3), col_index = 1:3)
}
