% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdata.R
\name{gdata}
\alias{gdata}
\title{Generate Simulation Data}
\usage{
gdata(tau, n, d, rho, fun_list, col_index)
}
\arguments{
\item{tau}{a number,which is used to generate unif(0, tau) random number.}

\item{n}{the number of observations you want to generate.}

\item{d}{the number of covariates you want to generate.}

\item{rho}{the number to control the covariance between two adjacent columns of generated data.}

\item{fun_list}{a list, whose elements are functions to be applied to covariates specified by \code{col_index} one by one.}

\item{col_index}{a vector, which specifies the indexes of covariates corresponding to \code{fun_list} in the data matrix.}
}
\value{
A list containing:\tabular{ll}{
\code{y} \tab A numeric vector recording logarithm survival time. \cr
\tab \cr
\code{x} \tab The covariates matrix. \cr
\tab \cr
\code{delta} \tab A numeric vector that consists of status indicator, normally 0=alive, 1=dead. \cr
}
}
\description{
This function generates simulation survival data.
}
\examples{
f1 <- function(x) return(x)
f2 <- function(x) return(-x)
f3 <- function(x) return(x^2)
gdata(tau = 1, n = 200, d = 30, rho = 0.25,
      fun_list = list(f1, f2, f3), col_index = 1:3)
}
