% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMweight.R
\name{KMweight}
\alias{KMweight}
\title{Kaplan-Meier Weights.}
\usage{
KMweight(data)
}
\arguments{
\item{data}{a list, in which \code{x} is the covariates matrix, \code{y} is the logarithm survival time, and \code{delta} is the status indicator, normally 0=alive, 1=dead.}
}
\value{
A vector denotes Kaplan-Meier weights vector.
}
\description{
This function is used to calculate Kaplan-Meier weights vector.
}
\examples{
f1 <- function(x) return(x)
f2 <- function(x) return(-x)
f3 <- function(x) return(x^2)
mydata <- gdata(tau = 1, n = 200, d = 30, rho = 0.25,
                fun_list = list(f1, f2, f3), col_index = 1:3)
KMweight(mydata)
}
