% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l1_norm}
\alias{l1_norm}
\title{Calculate \eqn{L1}-norm}
\usage{
l1_norm(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
the \eqn{L_1}-norm of x.
}
\description{
This function calculates the \eqn{L_1}-norm of a vector.
}
