% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l2_norm}
\alias{l2_norm}
\title{Calculate \eqn{L2}-norm}
\usage{
l2_norm(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
the \eqn{L_2}-norm of x.
}
\description{
This function calculates the \eqn{L_2}-norm of a vector.
}
