% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection}
\alias{selection}
\title{Select the Optimal Estimator}
\usage{
selection(theta.mat, n, Loss)
}
\arguments{
\item{theta.mat}{a three-dimensional array of estimated values of coefficients under each pair of tuning parameters.
The structure refers to \code{Details}}

\item{n}{the number of observations.}

\item{Loss}{a matrix containing the value of loss function under each pair of tuning parameters.The structure refers to \code{Details}}
}
\value{
The optimal estimator under the EBIC.
}
\description{
This function selects the optimal estimator under the EBIC criterion.
}
\details{
\code{theta[i,j,]} contains the estimator under the \eqn{ith} value of the first tuning parameter and
the \eqn{jth} value of the second parameter.\code{Loss[i,j]} is the weighted least squares loss function value when the estimator is \code{theta[i,j,]}.
The EBIC is extended Bayesian information criterion proposed by Chen and Chen
(2008) combining the ideas suggested by Gray (1992) to calculate the degree of freedom.{TODO}
}
\note{
See the example of \code{\link{BMD}} function for an example.
}
