% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimuData.R
\docType{data}
\name{SimuData}
\alias{SimuData}
\title{Simulation Data}
\format{
A data frame with 200 rows and 210 columns:
\describe{
\item{row}{observation}
\item{column}{every 7 columns are the orthogonal b-splines expansion of one variable}
}
}
\usage{
SimuData
}
\description{
A dataset containing the 200 obversations and 30 variables with each variable being expanded to 7 orthogonal b-splines.
}
\note{
This dataset is used to run the example of \code{\link{BMD}} function,
and it is generated by \code{\link{gdata}} function with \eqn{f_1 = 3.7x,f2 = -4.5x} and \eqn{f_3 = 5\sin x+10 cos 2x}.
}
\keyword{datasets}
