% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sortxy}
\alias{sortxy}
\title{Sorted Estimation}
\usage{
sortxy(x, y, pointx)
}
\arguments{
\item{x}{a vector consisting of the covariate values.}

\item{y}{a vector consisting of the estimated response values corresponding to \code{x}.}

\item{pointx}{a vector consisting of the values of the covariates to be predicted.}
}
\value{
the sorted estimation at \code{pointx}.
}
\description{
This function returns the sorted estimation at \code{pointx}.
}
\details{
Denote \eqn{x_{(i)}} as the \eqn{ith} order statistics of \eqn{x},
and \eqn{y_{(i)}} is \eqn{y} value corresponding to \eqn{x_{(i)}}.
The estimator of \eqn{y} value of \eqn{pointx_j} is \eqn{y_{(i)}} if
\eqn{x_{(i)}\leq pointx_j < x_{(i+1)}}. If \eqn{x_{(n)}\leq pointx_j},
then the estimator is \eqn{y_{(n)}} and if \eqn{pointx_j < x_{(1)}} then the
estimator is \eqn{y_{(1)}}.
}
\examples{
sortxy(x = rnorm(100), y = rnorm(100), pointx = rnorm(100))
}
